refactor(spans): extract shared OTLP anyValueToPlain helper
Replace three near-identical AnyValue->plain-JS flatteners (the transform's
attr_string capture, the metadata enricher, the GenAI content parser) with one
shared helper in otlp/any-value.ts. The shared version captures a `__proto__`
key as an own property instead of silently hijacking the object prototype.

Addresses review feedback on #4035.
